requirements.txt
langchain-openai==1.1.6
langchain-community==0.4.1
langgraph==1.0.3
sqlalchemy==2.0.46
psycopg2-binary==2.9.10
streamlit==1.55.0
matplotlib==3.10.8
squarify==0.4.4
python-dotenv==1.2.1
